+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## Upcoming
 * Sped up hammers forging speed
+* 20 stamina are now consumed when hitting things
 
 ## 3.11.1
 * Fishing is now slowed down based off season